requirements.txt
Django==1.10.3
Markdown==2.6.7
Pygments==2.1.3
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
cffi==1.9.1
decorator==4.0.10
django-cors-headers==1.3.1
django-filter==1.0.0
djangorestframework==3.5.3
enum34==1.1.6
idna==2.1
ipaddress==1.0.17
ipython==5.1.0
ipython-genutils==0.1.0
lxml==3.6.4
mysqlclient==1.3.9
ndg-httpsclient==0.4.2
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
ptyprocess==0.5.1
pyOpenSSL==16.2.0
pyasn1==0.1.9
pycparser==2.17
requests==2.12.3
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
wcwidth==0.1.7
wsgiref==0.1.2
